augur: share augur_config fixture, unify minimal_config builder
Pull the duplicated `augur_config` fixture up into a single `augur/conftest.py`
that both the api and product test trees inherit — the per-package conftests
depend on `//augur:conftest` so pytest finds it in their runfiles — instead of
each conftest defining its own copy.

Fold the two near-identical `_minimal_config()` builders (config_test and
portfolio_sources_test) into one `minimal_config()` in `augur/api/testing.py`,
on a single `owner` agent. config_test's `alpha` references become `owner`, and
the one snapshot assertion that pinned the old per-file default is updated.

BAZEL_TEST_INVOCATIONS=buildbuddy:2f19a66b-424e-4ea4-9b08-f26a48b80169
