+++ augur/product/conftest.py
@@ -7,7 +7,6 @@ import pytest
 from augur.api.catalog import build_catalog
 from augur.api.config import Config
 from augur.api.portfolio_sources import resolve_portfolio_sources
-from augur.api.testing import load_fixture_config
 from augur.api.wire import CatalogResponse
 from augur.model.exogenous import Sampler
 from augur.product.scenarios import resolve_primary_agent_id, sim_locations_from_config
@@ -15,11 +14,6 @@ from augur.product.service import ProductService
 
 # What the `make_product_service` fixture hands tests: build a ProductService for one model.
 MakeProductService = Callable[..., ProductService]
-
-
-@pytest.fixture(scope="module")
-def augur_config() -> Config:
-    return load_fixture_config()
 
 
 @pytest.fixture(scope="module")
